refactor: 자동차 이름 가져오기 기능 module 분리
